pyproject.toml
[project]
name = "newsrag"
version = "0.1.0"
description = "Agentic AI-powered news aggregator"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.2",
    "beautifulsoup4>=4.12.0",
    "websockets>=11.0.3",
    "pymupdf>=1.23.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "asyncio>=3.4.3",
    "langchain-groq>=0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
